fixed styling for navbar, card items.  Created handlebar page for boardpage in readonly view.  Changed form description for the claim button.
