+++ app/public/css/navbar.css
@@ -17,6 +17,24 @@
     width: 10%;
     
   }
+  .dropdown .btn-primary {
+    background-color: transparent;
+    border-style: none;
+    padding: 0;
+    position: relative;
+    top: 3px;
+  }
+  .dropdown-menu {
+    width: 110%;
+    left:-20px;
+    border-radius: 0;
+    background-color: #4EB266;
+    box-shadow:none;
+  }
+  .dropdown-menu li a {
+    font-size: 14pt;
+    color:white;
+  }
 
 }
 
@@ -25,9 +43,6 @@
     background-color: transparent;
     border-style: none;
     padding: 0;
-  }
-  .dropdown-menu {
-    width: 100%;
   }
 
 .navbar-brand .navbar-foodboard {
