+++ app/views/boardpagero_home.handlebars
@@ -1,0 +1,224 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <Title>FoodBoard</Title>
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+
+
+  <link rel="stylesheet" href="/public/css/home.css">
+  <link rel="stylesheet" href="/public/css/login-register.css">
+
+  <script src="/public/js/index.js"></script>
+
+  <script type="text/javascript">
+    $(window).on('load', function () {
+      $('#logreg-modal').modal('show');
+    });
+  </script>
+
+
+  <script src="/socket.io/socket.io.js"></script>
+
+</head>
+
+<body id="page-top">
+  {{> navbar}}
+
+
+  <!-- Homepage-->
+  <div id="homepage-body-container" class="container-fluid">
+    <div id="background-lg" class="container-fluid-custom homepage-background">
+
+      <img class="background" src="./Pictures/hero-image-flat.jpg" alt="background">
+
+    </div>
+    <!-- background image for mobile -->
+    <div id="background-sm" class="container-fluid-custom homepage-background-sm">
+      <img class="background" src="./Pictures/hero-image.jpg" alt="background">
+    </div>
+
+    <div id="homepage-body" class="text-center">
+      <img class="img-responsive" src="./Pictures/largelogo-text-transparent.png" alt="FoodBoard logo">
+      <p>Share More, Waste Less</p>
+      <button type="button" class="btn btn-primary btn-lg" data-toggle="modal" data-target="#logreg-modal">Get Started!</button>
+    </div>
+  </div>
+
+  <!-- How it works section-->
+  <div class="container-fluid container-fluid-custom" id="how-it-works">
+    <div class="jumbotron jumbotron-custom">
+      <div class="row">
+        <div class="col-lg-12 text-center">
+          <b>VALUES</b>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col-lg-4 text-center">
+          <img src="./Pictures/sustain.png" alt="Sustain">
+          <div>
+            <h2>
+              <b>Sustain</b>
+            </h2>
+            <p>Reduce the amount of perishable foods in your community.</p>
+          </div>
+        </div>
+        <div class="col-lg-4 text-center">
+          <img src="./Pictures/groceries.png" alt="Share">
+          <div>
+            <h2>
+              <b>Share</b>
+            </h2>
+            <p>Post and claim items with the click of a button.</p>
+          </div>
+        </div>
+        <div class="col-lg-4 text-center">
+          <img src="./Pictures/grow.png" alt="Grow">
+          <div>
+            <h2>
+              <b>Grow</b>
+            </h2>
+            <p>Create a sustainable community within your neighborhood.</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  </br>
+  </br>
+  </br>
+  </br>
+  </br>
+  </br>
+
+
+  <!-- OUR TEAM -->
+  <div class="container-fluid text-center" id="our-team">
+    <div class="row">
+      <div class="col-lg-12 text-center" id="our-team-title">
+        <b>OUR TEAM</b>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-lg-1"></div>
+      <div class="col-lg-2 text-center">
+        <img class="img-circle" src="./Pictures/HeadShots/Angela.gif" alt="Angela Yeung">
+        <h3>Angela Yeung</h3>
+        <p>Front-End Developer</p>
+      </div>
+      <div class="col-lg-2 text-center">
+        <img class="img-circle" src="./Pictures/HeadShots/Derek.gif" alt="Derek Lee">
+        <h3>Derek Lee</h3>
+        <p>Operations Officer, Lead Test</p>
+      </div>
+      <div class="col-lg-2 text-center">
+        <img class="img-circle" src="./Pictures/HeadShots/James.gif" alt="James Nguyen">
+        <h3>James Nguyen</h3>
+        <p>Creative Officer</p>
+      </div>
+      <div class="col-lg-2 text-center">
+        <img class="img-circle" src="./Pictures/HeadShots/Michael.gif" alt="Michael Yu">
+        <h3>Michael Yu</h3>
+        <p>Back-End Developer</p>
+      </div>
+      <div class="col-lg-2 text-center">
+        <img class="img-circle" src="./Pictures/HeadShots/Robert.gif" alt="Robert Ozdoba">
+        <h3>Robert Ozdoba</h3>
+        <p>Back-End Developer</p>
+      </div>
+      <div class="col-lg-1"></div>
+    </div>
+  </div>
+
+
+  <!-- Modal: Login/Register Forms -->
+  <div class="modal fade" id="logreg-modal" tabindex="1" role="dialog">
+    <div class="modal-dialog modal-dialog-centered tabbed-modal" role="document">
+      <div id="logo-above-modal">
+        <img src="./Pictures/fb-logo-transparent.png" alt="FoodBoard">
+      </div>
+      <!-- Content-->
+      <div class="modal-content">
+        <!-- Navigation tabs: login/register-->
+        <ul class="nav nav-tabs tabs-2 green" role="tablist">
+          <li class="nav-item active">
+            <a class="nav-link" data-toggle="tab" href="#login-panel" role="tab"> Login</a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" data-toggle="tab" href="#register-panel" role="tab"> Register</a>
+          </li>
+        </ul>
+
+        <!-- Tab Panels -->
+        <div class="tab-content">
+          <!-- Login Panel -->
+          <div class="tab-pane fade in show active" id="login-panel" role="tabpanel">
+            <!-- Body of Login Panel -->
+            <div class="modal-body">
+              <div>
+                <p class="modal-text">Welcome back to FoodBoard!</p>
+              </div>
+              <form name="login" class="form-horizontal" action="/user_login" method="POST">
+                <!-- Login form email input -->
+                <div class="form-group">
+                  <div class="col-sm-2 input-addon">
+                    <span class="glyphicon glyphicon-envelope"></span>
+                  </div>
+                  <div class="col-sm-9">
+                    <input value="{{email}}" class="form-control" type="email" placeholder="Enter Email Address" name="login_email" required>
+                  </div>
+                </div>
+                <!-- Login form password input -->
+                <div class="form-group">
+                  <div class="col-sm-2 input-addon">
+                    <span class="glyphicon glyphicon-lock"></span>
+                  </div>
+                  <div class="col-sm-9">
+                    <input class="form-control" type="password" placeholder="Enter Password" name="login_pwd" required>
+                  </div>
+                </div>
+                <div class="form-group">
+                  <button type="submit" class="btn btn-success center-block" value="Login">Login</button>
+                </div>
+              </form>
+            </div>
+          </div>
+          {{> register}}
+          <!-- Register Panel Footer -->
+          <!-- <div class="modal-footer">
+                <button type="submit" class="btn btn-success center-block" value="Register">Register</button>
+              </div> -->
+        </div>
+      </div>
+    </div>
+  </div>
+  </div>
+
+
+  <!-- FAQ Section-->
+  <!-- <div class="container-fluid" id="faq">
+    <div class="row">
+      <div class="text-center">FAQ</div>
+    </div>
+    <div class="row">
+      <div class="col-sm-4 text-center">
+        1.
+      </div>
+      <div class="col-sm-8">
+        <p class="question">The question will go here.</p>
+        <p class="answer">Answer to the question</p>
+      </div>
+    </div>
+  </div> -->
+
+</body>
+
+<footer>
+
+</footer>
+
+</html>
