+++ app/views/partials/navbar.handlebars
@@ -19,9 +19,7 @@
                     <a class="btn btn-primary btn-lg dropdown-toggle glyphicon glyphicon-menu-hamburger" data-toggle="dropdown"></a>
                     <ul class="dropdown-menu">
                         <li class="nav-item text-center">
-                            <form action="/board">
-                                <input type="submit" class="linkButton" value="BOARD" />
-                            </form>
+                            <a href="/boardpagero" class="nav-link">BOARD</a>
                         </li>
                         <li class="nav-item text-center">
                             <a href="#our-team" class="nav-link">OUR TEAM</a>
