+++ app/views/partials/navbar_ro.handlebars
@@ -1,0 +1,41 @@
+<header id="header">
+  <div class="container-fluid">
+    <nav class="navbar navbar-expand-sm navbar-fixed-top navbar-custom">
+      <div class="nav navbar-brand">
+        <a class="navbar-foodboard" href="./index.html">
+          <img src="./Pictures/fb-logo-transparent.png" alt="FoodBoard">
+        </a>
+      </div>
+
+      <!-- navbar logo for mobile-->
+      <div class="nav navbar-brand-sm">
+        <div class="text-center">
+
+          <a class="navbar-foodboard" href="#page-top">
+            <img src="./Pictures/fb-logo-transparent.png" alt="FoodBoard">
+          </a>
+        </div>
+        <div class="navbar-right dropdown">
+          <a class="btn btn-primary btn-lg dropdown-toggle glyphicon glyphicon-menu-hamburger" data-toggle="dropdown"></a>
+          <ul class="dropdown-menu">
+            <li class="nav-item text-center">
+              <a href="/" class="nav-link" id="how-link">HOME</a>
+            </li>
+            <li class="nav-item text-center">
+              <a href="/boardpagero_home" class="nav-link">LOGIN</a>
+            </li>
+          </ul>
+        </div>
+      </div>
+      <!--Nav bar for Desktop-->
+      <ul class="nav navbar-nav navbar-right navbar-right-custom">
+        <li class="nav-item">
+          <a href="/" class="nav-link" id="how-link">HOME</a>
+        </li>
+        <li class="nav-item text-center">
+          <a href="/boardpagero_home" class="nav-link">LOGIN</a>
+        </li>
+      </ul>
+    </nav>
+  </div>
+</header>
