+++ app/views/partials/navbar_success.handlebars
@@ -16,7 +16,11 @@
                     </a>
                 </div>
                 <div class="navbar-right dropdown">
+
                     <a class="btn btn-primary btn-lg dropdown-toggle glyphicon glyphicon-menu-hamburger" data-toggle="dropdown"></a>
+                    <button type="button" class="btn btn-default" aria-label="Right Align" class="btn additem" data-toggle="modal" data-target="#itemModal">
+                        <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
+                    </button>
                     <ul class="dropdown-menu">
                         <li class="nav-item text-center">
                             <a href="/logout" class="nav-link">LOGOUT</a>
